Changed Project Structure to contain UI classes in one shared folder
