+++ src/ui/egui_textures.rs
@@ -1,7 +1,8 @@
 use std::collections::HashMap;
 use bevy::prelude::*;
 use bevy_egui::{egui, EguiContext};
-use bevy_egui::egui::TextureId;
+use bevy_egui::egui::FontFamily::Proportional;
+use bevy_egui::egui::{FontId, TextureId};
 use libexodus::player::Player;
 use libexodus::tiles::{AtlasIndex, Tile};
 use crate::{CurrentMapTextureAtlasHandle, CurrentPlayerTextureAtlasHandle};
@@ -66,4 +67,32 @@ pub fn atlas_to_egui_textures(
         textures,
         player_textures: textures_p,
     });
+}
+
+pub fn egui_fonts(ctx: &egui::Context) -> () {
+    let mut fonts = egui::FontDefinitions::default();
+    fonts.font_data.insert("exodus".to_owned(),
+                           egui::FontData::from_static(include_bytes!("../../assets/fonts/PublicPixel.ttf")));
+    fonts.families
+        .entry(egui::FontFamily::Proportional)
+        .or_default()
+        .insert(0, "exodus".to_owned());
+    fonts.families
+        .entry(egui::FontFamily::Monospace)
+        .or_default()
+        .push("exodus".to_owned());
+    ctx.set_fonts(fonts);
+
+    let mut style = (*ctx.style()).clone();
+    style.text_styles = [
+        (egui::TextStyle::Heading, FontId::new(30.0, Proportional)),
+        (egui::TextStyle::Name("Heading2".into()), FontId::new(25.0, Proportional)),
+        (egui::TextStyle::Name("Context".into()), FontId::new(23.0, Proportional)),
+        (egui::TextStyle::Body, FontId::new(18.0, Proportional)),
+        (egui::TextStyle::Monospace, FontId::new(16.0, Proportional)),
+        (egui::TextStyle::Button, FontId::new(20.0, Proportional)),
+        (egui::TextStyle::Small, FontId::new(10.0, Proportional)),
+    ]
+        .into();
+    ctx.set_style(style);
 }
